requirements.txt
# Core
google-adk>=1.0.0
google-cloud-aiplatform>=1.52.0
google-cloud-healthcare
google-cloud-firestore>=2.16.0
google-cloud-pubsub>=2.21.0
google-cloud-dlp>=3.15.0
google-cloud-secret-manager>=2.20.0
google-generativeai

# LangGraph
langgraph>=1.1.0
langchain-core>=1.2.0
langsmith>=0.7.0
langgraph-cli[inmem]

# Utilities
aiohttp>=3.9.0
pydantic>=2.7.0
python-dotenv>=1.0.0
nest_asyncio>=1.6.0
protobuf>=5.29.0
